Add dqn_cartpole_tch (#36)

+++ border-tch-agent/src/dqn/base.rs
@@ -229,8 +229,8 @@ where
                 }
             } else {
                 if fastrand::f32() < 0.01 {
-                    let n_actions = a.size()[1] as i32;
-                    let a = fastrand::i32(0..n_actions);
+                    let n_actions = a.size()[1] as i64;
+                    let a = fastrand::i64(0..n_actions);
                     Tensor::from(a)
                 } else {
                     a.argmax(-1, true)
